move base websocket mixin into its own module
